fix(gig-db): improve exports structure
